requirements.txt
Babel==2.11.0
click==8.1.3
deprecation==2.1.0
docxcompose==1.3.7
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
packaging==21.3
pdfrw==0.4
pikepdf==6.2.5
Pillow==9.3.0
pyparsing==3.0.9
python-docx==0.8.11
pytz==2022.6
six==1.16.0
Werkzeug==2.2.2
